+++ Heavy Space/src/shared/functionality/Globals.java
@@ -3,6 +3,6 @@ package shared.functionality;
 public class Globals {
 	public static long now;
 	public static float dt;
-	public static short tick;
+	public static int tick;
 	public static short snapshotTick;
 }
